backend/lukete/requirements.txt
psycopg2-binary==2.9.11
pandas==2.3.3
streamlit==1.54.0
python-dotenv==1.0.1
